requirements.txt
customtkinter
pandas
packaging
pynput
Pillow